+++ leia-models/src/test/java/com/grookage/leia/models/request/SearchRequestTest.java
@@ -29,5 +29,15 @@ class SearchRequestTest {
         final var namespaceRequest = ResourceHelper.getResource("request/searchRequest.json", SearchRequest.class);
         Assertions.assertNotNull(namespaceRequest);
         Assertions.assertTrue(namespaceRequest.getNamespaces().contains("test"));
+        Assertions.assertTrue(namespaceRequest.getOrgIds().contains("testOrgId"));
+    }
+
+    @Test
+    @SneakyThrows
+    void testNamespaceRequestAlias() {
+        final var namespaceRequest = ResourceHelper.getResource("request/searchRequestAlias.json", SearchRequest.class);
+        Assertions.assertNotNull(namespaceRequest);
+        Assertions.assertTrue(namespaceRequest.getNamespaces().contains("test"));
+        Assertions.assertTrue(namespaceRequest.getOrgIds().contains("testOrg"));
     }
 }