+++ leia-repository/src/main/java/com/grookage/leia/repository/SchemaRepository.java
@@ -25,8 +25,6 @@ import java.util.Optional;
 
 public interface SchemaRepository {
 
-    boolean createdRecordExists(final SchemaKey schemaKey);
-
     void create(final SchemaDetails schema);
 
     void update(final SchemaDetails schema);